refactor files to match camelCase convention
